Updated Validations and removed launch.json file

+++ WHO-NOK/WHO.NOK.API/Controllers/UserController.cs
@@ -7,8 +7,10 @@
 namespace WHO.NOK.API.Controllers
 {
     using DocumentFormat.OpenXml.EMMA;
+    using FluentValidation;
     using Microsoft.AspNetCore.Authorization;
     using Microsoft.AspNetCore.Mvc;
+    using Microsoft.CodeAnalysis.CSharp.Syntax;
     using Microsoft.Graph.Models;
     using WHO.NOK.API.Helper;
     using WHO.NOK.BusinessLogic.Features.User;
@@ -19,6 +21,7 @@ namespace WHO.NOK.API.Controllers
     using WHO.NOK.BusinessLogic.ViewModels.PaginatedResponses;
     using WHO.NOK.BusinessLogic.ViewModels.User;
     using WHO.NOK.Core.Common;
+    using WHO.NOK.Core.ResponseMiddleware;
     using WHO.NOK.Core.Wrappers;
     using WHO.NOKS.BusinessLogic.Features.User;
 
@@ -31,13 +34,39 @@ namespace WHO.NOK.API.Controllers
     public class UserController : BaseController
     {
         private readonly IUserService userService;
+        private readonly IValidator<ActivateUserCommand> activateUserValidator;
+        private readonly IValidator<CreateUserCommand> createUserValidator;
+        private readonly IValidator<DeactivateUserCommand> deactivateUserValidator;
+        private readonly IValidator<DeleteUserCommand> deleteUserValidator;
+        private readonly IValidator<UpdateUserDetailsCommand> updateUserDetailsValidator;
+        private readonly IValidator<GetUsersCommand> getUsersValidator;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="UserController"/> class.
         /// </summary>
         /// <param name="userService"><see cref="IUserService"/> Service.</param>
-        public UserController(IUserService userService) {
+        /// <param name="activateUserCommand"> Validate <see cref="ActivateUserCommand"/> command. </param>
+        /// <param name="createUserCommand"> Validate <see cref="CreateUserCommand"/> command. </param>
+        /// <param name="deactivateUserCommand"> Validate <see cref="DeactivateUserCommand"/> command. </param>
+        /// <param name="deleteUserCommand"> Validate <see cref="DeleteUserCommand"/> command. </param>
+        /// <param name="updateUserDetailsCommand"> Validate <see cref="UpdateUserDetailsCommand"/> command. </param>
+        /// <param name="getUsersCommand"> Validate <see cref="GetUsersCommand"/> command. </param>
+        public UserController(
+            IUserService userService,
+            IValidator<ActivateUserCommand> activateUserCommand,
+            IValidator<CreateUserCommand> createUserCommand,
+            IValidator<DeactivateUserCommand> deactivateUserCommand,
+            IValidator<DeleteUserCommand> deleteUserCommand,
+            IValidator<UpdateUserDetailsCommand> updateUserDetailsCommand,
+            IValidator<GetUsersCommand> getUsersCommand)
+        {
             this.userService = userService;
+            this.activateUserValidator = activateUserCommand;
+            this.createUserValidator = createUserCommand;
+            this.deactivateUserValidator = deactivateUserCommand;
+            this.deleteUserValidator = deleteUserCommand;
+            this.updateUserDetailsValidator = updateUserDetailsCommand;
+            this.getUsersValidator = getUsersCommand;
         }
 
         /// <summary>
@@ -50,9 +79,10 @@ namespace WHO.NOK.API.Controllers
         [ProducesResponseType(typeof(ApiResponse<PaginatedResponse<UserViewModel>>), 200)]
         public async Task<ActionResult<PaginatedResponse<UserViewModel>>> GetUsers([FromBody] GetUsersCommand request)
         {
-            if (request == null)
+            var result = await this.getUsersValidator.ValidateAsync(request);
+            if (!result.IsValid)
             {
-                throw new ApiException("Request is null");
+                throw new ApiException(result.GetAllValidationErrors());
             }
 
             return this.Ok(await this.userService.GetAllPaginatedAndFilteredUsers(request));
@@ -66,16 +96,17 @@ namespace WHO.NOK.API.Controllers
         // POST: User
         [HttpPost("CreateUser")]
         [ProducesResponseType(typeof(ApiResponse<CurrentUserViewModel>), 200)]
-        public async Task<ActionResult<int>> CreateUser([FromForm] CreateUserCommand request)
+        public async Task<ActionResult<int>> CreateUser([FromBody] CreateUserCommand request)
         {
-            if (request == null)
+            var result = await this.createUserValidator.ValidateAsync(request);
+            if (!result.IsValid)
             {
-                throw new ApiException("Request is null");
+                throw new ApiException(result.GetAllValidationErrors());
             }
 
             return this.Ok(await this.userService.CreateUserAsync(request));
         }
-        
+
         /// <summary>
         /// Update the details of the user.
         /// </summary>
@@ -86,9 +117,10 @@ namespace WHO.NOK.API.Controllers
         [ProducesResponseType(typeof(ApiResponse<UserViewModel>), 200)]
         public async Task<IActionResult> UpdateUser([FromBody] UpdateUserDetailsCommand request)
         {
-            if (request == null)
+            var result = await this.updateUserDetailsValidator.ValidateAsync(request);
+            if (!result.IsValid)
             {
-                throw new ApiException("Request is null");
+                throw new ApiException(result.GetAllValidationErrors());
             }
 
             return this.Ok(await this.userService.UpdateUserDetailsAsync(request));
@@ -103,9 +135,10 @@ namespace WHO.NOK.API.Controllers
         [ProducesResponseType(typeof(ApiResponse<int>), 200)]
         public async Task<IActionResult> DeactivateUser([FromBody] DeactivateUserCommand request)
         {
-            if (request == null)
+            var result = await this.deactivateUserValidator.ValidateAsync(request);
+            if (!result.IsValid)
             {
-                throw new ApiException("Request is null");
+                throw new ApiException(result.GetAllValidationErrors());
             }
 
             return this.Ok(await this.userService.DeactivateUserAsync(request));
@@ -120,9 +153,10 @@ namespace WHO.NOK.API.Controllers
         [ProducesResponseType(typeof(ApiResponse<bool>), 200)]
         public async Task<IActionResult> ActivateUser([FromBody] ActivateUserCommand request)
         {
-            if (request == null)
+            var result = await this.activateUserValidator.ValidateAsync(request);
+            if (!result.IsValid)
             {
-                throw new ApiException("Request is null");
+                throw new ApiException(result.GetAllValidationErrors());
             }
 
             return this.Ok(await this.userService.ActivateUserAsync(request));
@@ -137,12 +171,11 @@ namespace WHO.NOK.API.Controllers
         [ProducesResponseType(typeof(ApiResponse<UserViewModel>), 200)]
         public async Task<IActionResult> DeleteUser(DeleteUserCommand request)
         {
-            // request.User = this.user;
-            // var result = await this.deleteUserValidator.ValidateAsync(request);
-            // if (!result.IsValid)
-            // {
-            //     throw new ApiException(result.GetAllValidationErrors());
-            // }
+            var result = await this.deleteUserValidator.ValidateAsync(request);
+            if (!result.IsValid)
+            {
+                throw new ApiException(result.GetAllValidationErrors());
+            }
 
             return this.Ok(await this.userService.DeleteUserAsync(request));
         }
